// src/components/ui/practice/itr/itr-7/components/itr-7-schedule-80ib.types.ts
export interface ITR7Schedule80IBData {
    schedule80IB: {
        section_80ib_9: { // a
            undertaking1: number;
            undertaking2: number;
        };
        section_80ib_10: { // b
            undertaking1: number;
            undertaking2: number;
        };
        section_80ib_11a: { // c
            undertaking1: number;
            undertaking2: number;
        };
        section_80ib_11a_integrated: { // d
            undertaking1: number;
            undertaking2: number;
        };
        total_deduction: number; // e
    };
}